Rename logging to _capture_stream
