+++ external/fv3util/fv3util/__init__.py
@@ -39,7 +39,7 @@ from .quantity import Quantity, QuantityMetadata
 from .units import ensure_equal_units, units_are_equal, UnitsError
 from .communicator import TileCommunicator, CubedSphereCommunicator, Communicator
 from ._xarray import to_dataset
-from ._logging import capture_stream
+from ._capture_stream import capture_stream
 from . import testing
 
 __version__ = "0.4.1"
