+++ external/fv3util/fv3util/_capture_stream.py
