+++ external/fv3util/tests/test__capture_stream.py
